requirements.txt
absl-py==0.2.2
asn1crypto==0.24.0
astor==0.7.1
astropy==3.0.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.5.3
cov-core==1.15.0
coverage==4.5.1
cryptography>=2.3
cycler==0.10.0
dask==0.18.1
decorator==4.3.0
dill==0.2.8.2
docopt==0.6.2
docutils==0.14
eventlet==0.23.0
fundamentals==1.6.0
gast==0.2.0
greenlet==0.4.14
grpcio==1.13.0
h5py==2.8.0
idna==2.7
Keras==2.0.4
kiwisolver==1.0.1
lockfile==0.12.2
luigi==2.7.6
Markdown==2.6.11
matplotlib==2.2.2
multiprocess==0.70.6.1
mysqlclient==1.3.13
networkx==2.1
nose2==0.7.4
numpy==1.14.5
pandas==0.23.3
panstamps==0.5.1
Pillow==5.2.0
protobuf==3.6.0
psutil==5.4.6
pycparser==2.18
pyfits==3.5
PyMySQL==0.9.2
pyparsing==2.2.0
pyprof2calltree==1.4.3
python-daemon==2.1.2
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
PyYAML==3.13
requests==2.19.1
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
Theano==1.0.2
threadpool==1.3.2
toolz==0.9.0
tornado==4.5.3
unicodecsv==0.14.1
urllib3==1.23
Werkzeug==0.14.1
gkutils>=0.2.22
